Agent-sim-module/agent/requirements.txt
openai-agents
py-clob-client
httpx
pydantic>=2.0.0
python-dotenv>=1.0.0
